move getFileNameByUri from service to activity
