+++ app/src/main/java/ar/com/fsck/rugmi/UploadService.java
@@ -17,11 +17,9 @@ import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.database.Cursor;
 import android.net.Uri;
 import android.os.Build;
 import android.os.IBinder;
-import android.provider.MediaStore;
 import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationManagerCompat;
 
@@ -35,37 +33,6 @@ public class UploadService extends IntentService {
         super("UploadService");
     }
 
-    public String getFileNameByUri(Uri uri) {
-        String fileName = "unknown"; //default fileName
-
-        if (uri == null) {
-            return fileName;
-        }
-
-        Uri filePathUri = uri;
-        if (uri.getScheme().toString().compareTo("content") == 0) {
-            Cursor cursor = null;
-            try {
-                cursor = getContentResolver().query(uri, null, null, null,
-                        null);
-            } catch (SecurityException e) {
-                return uri.getLastPathSegment();
-            }
-            if (cursor != null && cursor.moveToFirst()) {
-                // cursor contains three columns, _display_name, _size and _data
-                // no idea why we used to get _data, but it's often null
-                int column_index = cursor
-                        .getColumnIndexOrThrow(MediaStore.Images.Media.DISPLAY_NAME);
-                filePathUri = Uri.parse(cursor.getString(column_index));
-                return filePathUri.getLastPathSegment().toString();
-            }
-        } else if (uri.getScheme().compareTo("file") == 0) {
-            return filePathUri.getLastPathSegment().toString();
-        } else {
-            return fileName + "_" + filePathUri.getLastPathSegment();
-        }
-        return fileName;
-    }
 
     // see http://androidsnippets.com/multipart-http-requests
     public String multipartRequest(String urlTo, String[] posts,
